xfrmi: fix adding xfrm interface ip address with ipv6
The code actual checks for IFA_LOCAL and IFA_LABEL when searching for an existing ip address. This doesn't work with ipv6, because the kernel only returns IFA_ADDRESS for ipv6.
Since the xfrm interface is always a broadcast interface, we can always use IFA_ADDRESS for both ipv4 and ipv6. IFA_LABEL is not necessary because you can filter on the interface index.