ikev2: for AEAD create the IKE SA's cipher_context once
In calc_v2_keymat(), create the cipher_context and then
save it in the IKE SA.

This should mean that the IKEv2 AEAD code has been updated to
latest FIPS requirements.  It should also result in a tiny
performance improvement.

see #1535 in the GCM code, replace PK11_Encrypt() with PK11_AEADOp()