NAT: replace global keep-alive timer with one per-state
The global keep-alive timer was implemented by scanning all
states for anything behind a NAT.  Since this is expensive,
the code tried to avoid the scan by only scheduling when
needed.  Except it instead ended up not scheduling, even though
it was needed.

This dumps that code and instead schedules a per-state keep-alive
when the state is behind the NAT and keep-alive is enabled.

fix #1763 no keep-alive scheduled