ikev1: in quick mode, on mode-config server, try to recover existing lease
For instance:

- client initiates connection and is assigned mode-config lease
- client goes to sleep
- server boots connection because DPD fails
- client wakes and re-establishes (presulably to rekey)
  skipping mode-config as it things it still has lease

If, during Quick Mode, the server finds that the lease is
still available and last used by the client then it is recovered.
Otherwise, in an attempt to convince the client to restart from
scratch, the ISAKMP SA is blown away.

Note that this doesn't help when the server restarts (case when
lease was never assigned).  The client really needs to perform
a mode-config to acquire the lease and get current values of
other parameters such as the DNS server.