ipsec interface: save the connection's ipsec-interface address
When adding connection's address to the ipsec-interface,
save the reference in struct connection .ipsec_interface_address.
Use this, and not a lookup, when releasing the address.

Replace code in del_ipsec_interface() freeing the address
list with the pexpect()ation that no addresses are left -
the interface should only be released after all connections
have been unrouted (releasing the address) and disoriented.

(I suspect ipsec-interface isn't so good when its physical
interface disappears; with this it should only pexpect)

close #1579 free_ipsec_interface_address_list()
it's been deleted