ikev1: set ST from the CHILD or IKE SA after jumbo message switch
The jumbo message switch first tries to find the IKE (ISAKMP) SA for
the exchange (failure is an option), and then when Quick mode, also
tries to find the Child (IPsec) SA (but only when there's an IKE SA).
Then, after the switch, ST is set to the Child, failing that, IKE, and
failing that, NULL.

This way ST has a clearly defined value and, provided ST is non-NULL,
the existance of the IKE (ISAKMP) SA is guarenteed.

(This also means that the jumbo message switch no longer uses the
SEND_NOTIFICATION() or LOG_MESSAGE() hacks.  Same cannot be said of
code that follows though, sigh.)