iptfs: in add_sec_label_kernel_policy() don't go fishing for a Child SA
Officially bunkem.

With Labeled IPsec, the IKE SA is the one installing kernel
policy and is doing so _before_ there are children.  Hence
it can only use parameters from the config file.

This points to a second problem.  The kernel should be told
what to install - TUNNEL, TRANSPORT, IPTFS - using the mode
and not using a bonus iptfs bit.  See:
   pass KERNEL_MODE_IPTFS into kernel, not iptfs=true+tunnel #1964