addconn: do not load connections that were not requested
'addconn' program reads the entire ipsec.conf and parses and checks
every connection in it, even if asked to add just one.  This takes a
lot of time when ipsec.conf contains a lot of connections.

For example, it takes 30 minutes to load 1000 connections one by one
by calling 'ipsec add' for each of them separately.

Let's read the config, but only fully parse and validate connections
that user is actually interested in.  We still need to read the whole
file / hierarchy of files, but we can skip loading of connections that
are not specified on the command line.

Tested with an ipsec.conf containing 1000 connections with left
protoport=udp and the right protoport=udp/6081 or udp/geneve.  The
'ipsec add <conn>' is called for every connection separately one by
one, so 1000 calls in total.  The table below shows how much time
it takes to add all the connections.

    protoport      Before     After     SpeedUp
    -------------------------------------------
    udp/6081      29.6 min   24.5 sec     72x
    udp/geneve    22.9 min   23.9 sec     57x

Note: 'geneve' is faster than '6081', because getservbyname() doesn't
have to parse the whole /etc/services, only about 2/3 of it.

I'm not particularly happy with recursive search for aliases, but I
didn't find a better way to do that.  Alternative might be to just not
optimize the cases where 'also' is present.

Note: The main function in addconn.c still needs to check if the
names match the connections for the case the listings are requested.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>