testing f41: run crypto-policies test on the fedora domain
i... it's testing fedora's crypto-policies.  Also rename
test to addconn-23-fedora-crypto-policies.

As a bonus, since the fedora domain is already running f41,
the output can be updated without needing to wait for
the linux upgrade.  And as a double bonus, same thing
will happen when fedora advances to f42.

close #2063 f41: is addconn-23-system-crypto-policies really fedora-check-system-policies