CRLs: merge submit_crl_fetch_requests() into add_crl_fetch_request()
This de-tunes the CRL code slightly.

Instead of accumulating a list of CRLs and adding them as a batch,
the code adds each request individually.  This means that
each request requires a lock/unlock.  Since the fetch thread
doesn't hold a lock while processing a request, this isn't an issue.