testing: add exhaustive WIP newoe-30-shunt-slash{24,32}
these exhaustively test all combinations of:
  - negotation shunt: default, pass, hold
  - failure shunt: default, pass, hold, none
  - group template with /32 and /24 trap
  - good, broken IKE and broken Child
(it turns out this was a huge gap in testing)

Instead of playing with cops, pocs, blocks, clears, and privates,
this uses whack on road, and a normal conn on east.

close #2125 test coverage for negotation={,drop,pass} failure={,none,drop,pass} with slash{24,32}
ref #2108 Assertion failure with libreswan 5.1 with null authentication

Also suspect this made some of the existing OE tests redundant.