routing: rename replace_ipsec_with_bare_kernel_policy() ...
to uninstall_ipsec_kernel_policy().  Since the kernel policy has
a connection it is not bare.

In the past, any kernel policy with no state, was called
a bare shunt.  Since 5.0, only kernel polices with no state
and no connection are bare.  Further, they are only
created when an OE connection fails and the OE connection
instance is deleted.

Perhaps bare_shunt should be renamed to orphaned_kernel_policy.