ipsec down: don't delete IKE SA when it has policy +UP
For instance, a Child SA #2 for connection "A", has IKE SA #1 from
connection "B" and and both "A" and "B" are +UP.

`ipsec down` should take down the Child SA, but not the IKE SA
even though the IKE SA has no other children - can't argue
with the +UP bit.

Also revamp log messages: one line for delete; one line with
all the reason for keeping the IKE SA up.

see #2310 don't down the other sole parent when it's +UP