ikev2: when no subnet=, use subnet=${host} to search for template matching TS
When a connection does not specify {left,right}subnet=, the subnet
(selector) is defaulted to the host address converted to a selector.

When searching for a better template connection matching TS, and a
template end does not specify subnet=, use the template connection's
end's host address (converted) as the selector.

see #2380 IKEv2: Multiple host to subnet tunnels to same responder can result in TS_UNACCEPTABLE