ikev2: reassembled packet containing unencrypted payloads
Critically, the Next Payload field, when embedded in the
unencrypted payload immediately before the SKF payload is
changed to SK (as IKE_INTERMEDIATE expects).  For consistency
other fields such as length are also adjusted.

The reconstructed message replaces the old message in .payload
(.v2_sk_payload is dropped).

All the fields, such as .digest[] and .message_pbs that were
pointing into the old mesage are switched to the new reconstructed
message.