ikev2: detect and resolve simultaneous IKE_AUTH exchanges
When processing an IKE_AUTH request as responder, detect if there's
a simultaneous IKE SA on the same permanent connection where we are
the initiator with an outstanding IKE_AUTH request.

If this is detected, keep only one IKE SA based on this logic:

  - If the simultaneous (initiator) IKE SA already completed IKE_AUTH
    within the last second, keep it and reject the current (responder)
    IKE_AUTH request.

  - Otherwise, compare the initiator nonces from both exchanges and
    keep the IKE SA with the higher nonce value.

If rejecting responder's IKE SA we send AUTHENTICATION_FAILED. Otherwise,
we terminate the initiator's IKE SA locally. This is enabled by default
and can be disabled by connection option reject-simultaneous-ike-auth=no.

Signed-off-by: Ondrej Moris <omoris@redhat.com>
Signed-off-by: Daiki Ueno <dueno@redhat.com>